Adapt to modern glibc
Modern glibc have dropped their inline function mapping, so they use the
"real" name as their symbol now.

This adapts the dce code to this.

The old int ver is still passed as a dummy parameter via the c++ class
layer for Fxstat/Fxstat64 but that can be dropped later.

Signed-off-by: Anders Martinsson <anders.martinsson@sikabo.com>